Revert "chore: rename OrderItemButton -> ~Group"
This reverts commit 9f48fc752c3d1b64e30a89d271bbad80709bcd72.
